+++ core/src/lib.rs
@@ -1,14 +1,15 @@
 #![recursion_limit = "256"]
+#![cfg_attr(feature = "diagnostics", feature(proc_macro_diagnostic))]
 
 #[macro_use]
 extern crate quote;
-
 #[macro_use]
 extern crate syn;
-extern crate proc_macro2;
-
 extern crate fnv;
 extern crate ident_case;
+#[cfg(feature = "diagnostics")]
+extern crate proc_macro;
+extern crate proc_macro2;
 
 #[macro_use]
 mod macros_private;