+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.8.4
+- Use `syn::Error` to provide precise errors before `proc_macro::Diagnostic` is available
+- Add `diagnostics` feature flag to toggle between stable and unstable error backends
+- Attach error information in more contexts
+
 ## v0.8.3 (January 21, 2019)
 - Attach spans to errors in generated trait impls [#37](https://github.com/darling/issues/37)
 - Attach spans to errors for types with provided bespoke implementations